+++ zarrs_metadata/CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: Move `v3::array::data_type::DataTypeSize` to the crate root
 - **Breaking**: Rename `v2_to_v3::array_metadata_fill_value_v2_to_v3` to `fill_value_metadata_v2_to_v3`
 - **Breaking**: Rename `v2_to_v3::data_type_metadata_v2_to_v3_data_type` to `data_type_metadata_v2_to_v3`
+- `zfpy` codec: add support for reversible in case it gets fixed in `numcodecs`
 
 ### Removed
 - **Breaking**: Remove `fill_value::{HexString,FillValueFloat,FillValueFloatStringNonFinite}`
@@ -50,6 +51,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: Remove the `v2::array::codec` module which just contained re-exports of V3 codecs
 - **Breaking**: Remove `v2_to_v3::DataTypeMetadataV2UnsupportedDataTypeError`
 - **Breaking**: Remove the `Metadata` alias in the crate root (deprecated since 0.17)
+
+### Fixed
+- Stricter parameter validation for the `zfpy` Codec
 
 ## [0.3.7] - 2025-04-11